requirements.txt
PySide6>=6.6
openai>=1.0.0
llama-cpp-python>=0.2.0
